app/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database (SQLite for simplicity, can upgrade to PostgreSQL)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP Client for Model Servers
httpx==0.26.0
aiohttp==3.9.1

# Background Jobs - FIXED: Use redis version compatible with both rq and libretranslate
redis==4.4.4
rq==1.16.1

# AI/ML Integration
google-cloud-vision==3.5.0
google-generativeai==0.3.2

# Translation - We'll use LibreTranslate API, not the full package
# libretranslate==1.5.5  # Removed - causes conflicts
# Instead, we'll just use HTTP requests to LibreTranslate API

# Utils
python-dotenv==1.0.0
pillow==10.2.0
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3